src/main/java/com/github/xdshent/leetcode/array/MedianOfTwoSortedArraysSolution.java
package com.github.xdshent.leetcode.array;

/**
 * 4. Median of Two Sorted Arrays
 * There are two sorted arrays nums1 and nums2 of size m and n respectively.
 * Find the median of the two sorted arrays. The overall run time complexity should be O(log (m+n))
 * You may assume nums1 and nums2 cannot be both empty.
 * <p>
 * Example 1:
 * nums1 = [1, 3]
 * nums2 = [2]
 * The median is 2.0
 * <p>
 * Example 2:
 * nums1 = [1, 2]
 * nums2 = [3, 4]
 * <p>
 * The median is (2 + 3)/2 = 2.5
 *
 * @author xdshen
 */
public class MedianOfTwoSortedArraysSolution {

    /**
     * Time Complexity: O(log(M+N))
     * Space Complexity: O(1)
     *
     * @param nums1
     * @param nums2
     * @return
     */
    public double findMedianSortedArrays(int[] nums1, int[] nums2) {
        int index1 = 0;
        int index2 = 0;
        int med1 = 0;
        int med2 = 0;
        int minLength = 2;

        for (int i = 0; i <= (nums1.length + nums2.length) / minLength; i++) {
            med1 = med2;

            if (index1 == nums1.length) {
                med2 = nums2[index2];
                index2++;
            } else if (index2 == nums2.length) {
                med2 = nums1[index1];
                index1++;
            } else if (nums1[index1] < nums2[index2]) {
                med2 = nums1[index1];
                index1++;
            } else {
                med2 = nums2[index2];
                index2++;
            }
        }

        if ((nums1.length + nums2.length) % minLength == 0) {
            return (double) (med1 + med2) / minLength;
        }

        return med2;
    }
}
